requirements.txt
streamlit
PyPDF2
google-generativeai
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
python-dotenv
